Added extra security checks
git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@625 c91229c3-7414-0410-bfa2-8a42b809f60b

+++ presentation/lookAndFeel/knowledgeTree/documentmanagement/addDocumentBL.php
@@ -27,35 +27,59 @@ if (checkSession()) {
 		if (Permission::userHasFolderWritePermission($fFolderID)) {
 			//user has permission to add document to this folder
 			if (isset($fForStore)) {
-				//user wants to store a document
-				//create the document in the database
-				//var_dump($_FILES);
-				$oDocument = & PhysicalDocumentManager::createDocumentFromUploadedFile($_FILES['fFile'], $fFolderID);
-				if ($oDocument->create()) {
-					//if the document was successfully created in the db, then store it on the file system
-					if (PhysicalDocumentManager::uploadPhysicalDocument($oDocument, $fFolderID, "None", $_FILES['fFile']['tmp_name'])) {
-						redirect("$default->owl_root_url/control.php?action=viewDocument&fDocumentID=" . $oDocument->getID());
+				//user wants to store a document				
+				//make sure the user actually selected a file first
+				if (strlen($_FILES['fFile']['name']) > 0) {
+					//if the user selected a file to upload
+					//create the document in the database
+					$oDocument = & PhysicalDocumentManager::createDocumentFromUploadedFile($_FILES['fFile'], $fFolderID);
+					if (!(Document::documentExists($oDocument->getFileName(), $oDocument->getFolderID()))) {
+						if ($oDocument->create()) {
+							//if the document was successfully created in the db, then store it on the file system
+							if (PhysicalDocumentManager::uploadPhysicalDocument($oDocument, $fFolderID, "None", $_FILES['fFile']['tmp_name'])) {
+								redirect("$default->owl_root_url/control.php?action=viewDocument&fDocumentID=" . $oDocument->getID());
+							} else {
+								require_once("$default->owl_fs_root/presentation/webpageTemplate.inc");
+								$oDocument->delete();
+								$oPatternCustom = & new PatternCustom();
+								$oPatternCustom->setHtml(getBrowseAddPage($fFolderID));
+								$main->setCentralPayload($oPatternCustom);
+								$main->setFormAction($_SERVER["PHP_SELF"] . "?fFolderID=$fFolderID&fForStore=1");
+								$main->setFormEncType("multipart/form-data");
+								$main->setErrorMessage("An error occured while storing the document on the file system");
+								$main->render();
+							}
+						} else {
+							require_once("$default->owl_fs_root/presentation/webpageTemplate.inc");
+							$oPatternCustom = & new PatternCustom();
+							$oPatternCustom->setHtml(getBrowseAddPage($fFolderID));
+							$main->setCentralPayload($oPatternCustom);
+							$main->setFormAction($_SERVER["PHP_SELF"] . "?fFolderID=$fFolderID&fForStore=1");
+							$main->setFormEncType("multipart/form-data");
+							$main->setErrorMessage("An error occured while storing the document in the database");
+							$main->render();
+						}
 					} else {
-						require_once("$default->owl_fs_root/presentation/webpageTemplate.inc");
-						$oDocument->delete();
-						$oPatternCustom = & new PatternCustom();
-						$oPatternCustom->setHtml(getBrowseAddPage($fFolderID));
-						$main->setCentralPayload($oPatternCustom);
-						$main->setFormAction($_SERVER["PHP_SELF"] . "?fFolderID=$fFolderID&fForStore=1");
-						$main->setFormEncType("multipart/form-data");
-						$main->setErrorMessage("An error occured while storing the document on the file system");
-						$main->render();
+							require_once("$default->owl_fs_root/presentation/webpageTemplate.inc");
+							$oPatternCustom = & new PatternCustom();
+							$oPatternCustom->setHtml(getBrowseAddPage($fFolderID));
+							$main->setCentralPayload($oPatternCustom);
+							$main->setFormAction($_SERVER["PHP_SELF"] . "?fFolderID=$fFolderID&fForStore=1");
+							$main->setFormEncType("multipart/form-data");
+							$main->setErrorMessage("A document with this file name already exists in this folder");
+							$main->render();
 					}
 				} else {
 					require_once("$default->owl_fs_root/presentation/webpageTemplate.inc");
-					$oPatternCustom = & new PatternCustom();
-					$oPatternCustom->setHtml(getBrowseAddPage($fFolderID));
-					$main->setCentralPayload($oPatternCustom);
-					$main->setFormAction($_SERVER["PHP_SELF"] . "?fFolderID=$fFolderID&fForStore=1");
-					$main->setFormEncType("multipart/form-data");
-					$main->setErrorMessage("An error occured while storing the document in the database");
-					$main->render();
+							$oPatternCustom = & new PatternCustom();
+							$oPatternCustom->setHtml(getBrowseAddPage($fFolderID));
+							$main->setCentralPayload($oPatternCustom);
+							$main->setFormAction($_SERVER["PHP_SELF"] . "?fFolderID=$fFolderID&fForStore=1");
+							$main->setFormEncType("multipart/form-data");
+							$main->setErrorMessage("Please select a document by first clicking on 'Browse'.  Then click 'Add'");
+							$main->render();
 				}
+			
 			} else {
 				//we're still just browsing
 				require_once("$default->owl_fs_root/presentation/webpageTemplate.inc");
